+++ CHANGES.md
@@ -2,3 +2,4 @@
 - Removed Image Select Script (now it's dependant on EJO-base)
 - Added changelog
 - Temporarily disabled icon-picker
+- Allow image-size be overridden by theme